refactor(wow-core): split CommandWaitNotifier into separate files
- Move CommandWaitEndpoint and related classes to new file
- Create new file for ExtractedWaitStrategy and related functions
- Simplify CommandWaitNotifier class
