+++ wow-core/src/main/kotlin/me/ahoo/wow/command/wait/CommandWaitEndpoint.kt
@@ -1,0 +1,20 @@
+/*
+ * Copyright [2021-present] [ahoo wang <ahoowang@qq.com> (https://github.com/Ahoo-Wang)].
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package me.ahoo.wow.command.wait
+
+interface CommandWaitEndpoint {
+    val endpoint: String
+}
+
+data class SimpleCommandWaitEndpoint(override val endpoint: String) : CommandWaitEndpoint
